:basketball:

+++ lintcode/UniqueBinarySearchTree.java
@@ -38,11 +38,11 @@ public class UniqueBinarySearchTree {
     }
 
     public int numTrees(int n) {
-        int result = 1;
+        long result = 1;
         for (int i = 0; i < n; ++i) {
             result = result * 2 * (2 * i + 1) / (i + 2);
         }
-        return result;
+        return (int) result;
     }
 
 /*********************************************************/
